+++ pytorch_lightning/callbacks/model_checkpoint.py
@@ -22,15 +22,15 @@ Automatically save model checkpoints during training.
 
 import os
 import re
-from typing import Optional
+from typing import Any, Dict, Optional
 
 import numpy as np
 import torch
-
 from pytorch_lightning import _logger as log
 from pytorch_lightning.callbacks.base import Callback
-from pytorch_lightning.utilities import rank_zero_warn, rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only, rank_zero_warn
 from pytorch_lightning.utilities.cloud_io import get_filesystem
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
 class ModelCheckpoint(Callback):
@@ -47,11 +47,11 @@ class ModelCheckpoint(Callback):
             Example::
 
                 # custom path
-                # saves a file like: my/path/epoch_0.ckpt
+                # saves a file like: my/path/epoch=0.ckpt
                 >>> checkpoint_callback = ModelCheckpoint('my/path/')
 
                 # save any arbitrary metrics like `val_loss`, etc. in name
-                # saves a file like: my/path/epoch=2-val_loss=0.2_other_metric=0.3.ckpt
+                # saves a file like: my/path/epoch=2-val_loss=0.02-other_metric=0.03.ckpt
                 >>> checkpoint_callback = ModelCheckpoint(
                 ...     filepath='my/path/{epoch}-{val_loss:.2f}-{other_metric:.2f}'
                 ... )
@@ -97,9 +97,9 @@ class ModelCheckpoint(Callback):
         >>> trainer = Trainer(checkpoint_callback=checkpoint_callback)
 
         # save epoch and val_loss in name
-        # saves a file like: my/path/sample-mnist_epoch=02_val_loss=0.32.ckpt
+        # saves a file like: my/path/sample-mnist-epoch=02-val_loss=0.32.ckpt
         >>> checkpoint_callback = ModelCheckpoint(
-        ...     filepath='my/path/sample-mnist_{epoch:02d}-{val_loss:.2f}'
+        ...     filepath='my/path/sample-mnist-{epoch:02d}-{val_loss:.2f}'
         ... )
 
         # retrieve the best checkpoint after training
@@ -111,32 +111,43 @@ class ModelCheckpoint(Callback):
 
     """
 
-    CHECKPOINT_NAME_LAST = "last.ckpt"
+    CHECKPOINT_JOIN_CHAR = "-"
+    CHECKPOINT_NAME_LAST = "last"
     CHECKPOINT_STATE_BEST_SCORE = "checkpoint_callback_best_model_score"
     CHECKPOINT_STATE_BEST_PATH = "checkpoint_callback_best_model_path"
 
-    def __init__(self, filepath: Optional[str] = None, monitor: str = 'val_loss', verbose: bool = False,
-                 save_last: bool = False, save_top_k: int = 1, save_weights_only: bool = False,
-                 mode: str = 'auto', period: int = 1, prefix: str = ''):
+    def __init__(
+        self,
+        filepath: Optional[str] = None,
+        monitor: str = "checkpoint_on",
+        verbose: bool = False,
+        save_last: bool = False,
+        save_top_k: int = 1,
+        save_weights_only: bool = False,
+        mode: str = "auto",
+        period: int = 1,
+        prefix: str = "",
+    ):
         super().__init__()
-        if filepath:
-            self._fs = get_filesystem(filepath)
-        else:
-            self._fs = get_filesystem("")  # will give local fileystem
-        if save_top_k > 0 and filepath is not None and self._fs.isdir(filepath) and len(self._fs.ls(filepath)) > 0:
+        self._fs = get_filesystem(filepath if filepath is not None else "")
+        if (
+            save_top_k > 0
+            and filepath is not None
+            and self._fs.isdir(filepath)
+            and len(self._fs.ls(filepath)) > 0
+        ):
             rank_zero_warn(
                 f"Checkpoint directory {filepath} exists and is not empty with save_top_k != 0."
-                "All files in this directory will be deleted when a checkpoint is saved!"
+                " All files in this directory will be deleted when a checkpoint is saved!"
             )
-        self._rank = 0
 
         self.monitor = monitor
         self.verbose = verbose
-        if filepath is None:  # will be determined by trainer at runtime
+        if not filepath:  # will be determined by trainer at runtime
             self.dirpath, self.filename = None, None
         else:
             if self._fs.isdir(filepath):
-                self.dirpath, self.filename = filepath, "{epoch}"
+                self.dirpath, self.filename = filepath, None
             else:
                 if self._fs.protocol == "file":  # dont normalize remote paths
                     filepath = os.path.realpath(filepath)
@@ -150,44 +161,36 @@ class ModelCheckpoint(Callback):
         self.prefix = prefix
         self.best_k_models = {}
         # {filename: monitor}
-        self.kth_best_model_path = ''
+        self.kth_best_model_path = ""
         self.best_model_score = 0
-        self.best_model_path = ''
+        self.best_model_path = ""
+        self.last_model_path = ""
         self.save_function = None
         self.warned_result_obj = False
 
         torch_inf = torch.tensor(np.Inf)
         mode_dict = {
-            'min': (torch_inf, 'min'),
-            'max': (-torch_inf, 'max'),
-            'auto': (-torch_inf, 'max') if 'acc' in self.monitor or self.monitor.startswith('fmeasure')
-            else (torch_inf, 'min'),
+            "min": (torch_inf, "min"),
+            "max": (-torch_inf, "max"),
+            "auto": (-torch_inf, "max")
+            if "acc" in self.monitor or self.monitor.startswith("fmeasure")
+            else (torch_inf, "min"),
         }
 
         if mode not in mode_dict:
-            rank_zero_warn(f'ModelCheckpoint mode {mode} is unknown, '
-                           f'fallback to auto mode.', RuntimeWarning)
-            mode = 'auto'
+            rank_zero_warn(
+                f"ModelCheckpoint mode {mode} is unknown, " f"fallback to auto mode.",
+                RuntimeWarning,
+            )
+            mode = "auto"
 
         self.kth_value, self.mode = mode_dict[mode]
 
-    @property
-    def best(self):
-        rank_zero_warn("Attribute `best` has been renamed to `best_model_score` since v0.8.0"
-                       " and will be removed in v0.10.0", DeprecationWarning)
-        return self.best_model_score
-
-    @property
-    def kth_best_model(self):
-        rank_zero_warn("Attribute `kth_best_model` has been renamed to `kth_best_model_path` since v0.8.0"
-                       " and will be removed in v0.10.0", DeprecationWarning)
-        return self.kth_best_model_path
-
-    def _del_model(self, filepath):
+    def _del_model(self, filepath: str):
         if self._fs.exists(filepath):
             self._fs.rm(filepath)
 
-    def _save_model(self, filepath, trainer, pl_module):
+    def _save_model(self, filepath: str, trainer, pl_module):
 
         # in debugging, track when we save checkpoints
         trainer.dev_debugger.track_checkpointing_history(filepath)
@@ -201,26 +204,49 @@ class ModelCheckpoint(Callback):
         else:
             raise ValueError(".save_function() not set")
 
-    def check_monitor_top_k(self, current):
+    def check_monitor_top_k(self, current) -> bool:
         less_than_k_models = len(self.best_k_models) < self.save_top_k
         if less_than_k_models:
             return True
 
         if not isinstance(current, torch.Tensor):
             rank_zero_warn(
-                f'{current} is supposed to be a `torch.Tensor`. Saving checkpoint may not work correctly.'
-                f' HINT: check the value of {self.monitor} in your validation loop', RuntimeWarning
+                f"{current} is supposed to be a `torch.Tensor`. Saving checkpoint may not work correctly."
+                f" HINT: check the value of {self.monitor} in your validation loop",
+                RuntimeWarning,
             )
             current = torch.tensor(current)
 
-        monitor_op = {
-            "min": torch.lt,
-            "max": torch.gt,
-        }[self.mode]
+        monitor_op = {"min": torch.lt, "max": torch.gt}[self.mode]
 
         return monitor_op(current, self.best_k_models[self.kth_best_model_path])
 
-    def format_checkpoint_name(self, epoch, metrics, ver=None):
+    @classmethod
+    def _format_checkpoint_name(
+        cls,
+        filename: Optional[str],
+        epoch: int,
+        metrics: Dict[str, Any],
+        prefix: str = "",
+    ) -> str:
+        if not filename:
+            # filename is not set, use default name
+            filename = "{epoch}"
+        # check and parse user passed keys in the string
+        groups = re.findall(r"(\{.*?)[:\}]", filename)
+        if groups:
+            metrics["epoch"] = epoch
+            for group in groups:
+                name = group[1:]
+                filename = filename.replace(group, name + "={" + name)
+                if name not in metrics:
+                    metrics[name] = 0
+            filename = filename.format(**metrics)
+        return cls.CHECKPOINT_JOIN_CHAR.join([txt for txt in (prefix, filename) if txt])
+
+    def format_checkpoint_name(
+        self, epoch: int, metrics: Dict[str, Any], ver: Optional[int] = None
+    ) -> str:
         """Generate a filename according to the defined template.
 
         Example::
@@ -239,24 +265,13 @@ class ModelCheckpoint(Callback):
             >>> os.path.basename(ckpt.format_checkpoint_name(0, {}))
             'missing=0.ckpt'
         """
-        # check if user passed in keys to the string
-        groups = re.findall(r'(\{.*?)[:\}]', self.filename)
-
-        if len(groups) == 0:
-            # default name
-            filename = f'{self.prefix}_ckpt_epoch_{epoch}'
-        else:
-            metrics['epoch'] = epoch
-            filename = self.filename
-            for tmp in groups:
-                name = tmp[1:]
-                filename = filename.replace(tmp, name + '={' + name)
-                if name not in metrics:
-                    metrics[name] = 0
-            filename = filename.format(**metrics)
-        str_ver = f'_v{ver}' if ver is not None else ''
-        filepath = os.path.join(self.dirpath, self.prefix + filename + str_ver + '.ckpt')
-        return filepath
+        filename = self._format_checkpoint_name(
+            self.filename, epoch, metrics, prefix=self.prefix
+        )
+        if ver is not None:
+            filename = self.CHECKPOINT_JOIN_CHAR.join((filename, f"v{ver}"))
+        ckpt_name = f"{filename}.ckpt"
+        return os.path.join(self.dirpath, ckpt_name) if self.dirpath else ckpt_name
 
     @rank_zero_only
     def on_pretrain_routine_start(self, trainer, pl_module):
@@ -276,7 +291,7 @@ class ModelCheckpoint(Callback):
         if self.dirpath is not None:
             return  # short circuit
 
-        self.filename = '{epoch}'
+        self.filename = None
 
         if trainer.logger is not None:
             if trainer.weights_save_path != trainer.default_root_dir:
@@ -285,33 +300,23 @@ class ModelCheckpoint(Callback):
             else:
                 save_dir = trainer.logger.save_dir or trainer.default_root_dir
 
-            version = trainer.logger.version if isinstance(
-                trainer.logger.version, str) else f'version_{trainer.logger.version}'
+            version = (
+                trainer.logger.version
+                if isinstance(trainer.logger.version, str)
+                else f"version_{trainer.logger.version}"
+            )
             ckpt_path = os.path.join(
-                save_dir,
-                trainer.logger.name,
-                version,
-                "checkpoints"
+                save_dir, trainer.logger.name, version, "checkpoints"
             )
         else:
             ckpt_path = os.path.join(trainer.weights_save_path, "checkpoints")
 
         self.dirpath = ckpt_path
 
-        assert trainer.global_rank == 0, "tried to make a checkpoint from non global_rank=0"
+        assert (
+            trainer.global_rank == 0
+        ), "tried to make a checkpoint from non global_rank=0"
         self._fs.makedirs(self.dirpath, exist_ok=True)
-
-    def __warn_deprecated_monitor_key(self):
-        using_result_obj = os.environ.get('PL_USING_RESULT_OBJ', None)
-        invalid_key = self.monitor not in ['val_loss', 'checkpoint_on', 'loss', 'val_checkpoint_on']
-        if using_result_obj and not self.warned_result_obj and invalid_key:
-            self.warned_result_obj = True
-            m = f"""
-                    When using EvalResult(checkpoint_on=X) or TrainResult(checkpoint_on=X) the
-                    'monitor' key of ModelCheckpoint has no effect.
-                    Remove ModelCheckpoint(monitor='{self.monitor}) to fix')
-                """
-            rank_zero_warn(m)
 
     @rank_zero_only
     def on_validation_end(self, trainer, pl_module):
@@ -319,27 +324,28 @@ class ModelCheckpoint(Callback):
         if trainer.global_rank != 0:
             return
 
-        if trainer.running_sanity_check:
+        # no models are saved
+        if self.save_top_k == 0:
             return
 
-        # TODO: remove when dict results are deprecated
-        self.__warn_deprecated_monitor_key()
+        if trainer.running_sanity_check:
+            return
 
         metrics = trainer.logger_connector.callback_metrics
         epoch = trainer.current_epoch
 
-        # support structured results
-        if metrics.get('checkpoint_on') is not None:
-            self.monitor = 'checkpoint_on'
+        # validate metric
+        if not self._is_valid_monitor_key(metrics):
+            keys = list(metrics.keys())
+            m = f"""
+                ModelCheckpoint(monitor='{self.monitor}') not found in the returned metrics ({keys}),
+                "did you call result.log(f'{self.monitor}', tensor)?"""
+            raise MisconfigurationException(m)
 
-        # conditioned val metrics override conditioned train loop metrics
-        if metrics.get('val_checkpoint_on') is not None:
-            self.monitor = 'val_checkpoint_on'
-
-        if self.save_top_k == 0:
-            # no models are saved
-            return
-        if self.epoch_last_check is not None and (epoch - self.epoch_last_check) < self.period:
+        if (
+            self.epoch_last_check is not None
+            and (epoch - self.epoch_last_check) < self.period
+        ):
             # skipping in this term
             return
 
@@ -349,7 +355,9 @@ class ModelCheckpoint(Callback):
         filepath = self.format_checkpoint_name(epoch, ckpt_name_metrics)
         version_cnt = 0
         while self._fs.exists(filepath):
-            filepath = self.format_checkpoint_name(epoch, ckpt_name_metrics, ver=version_cnt)
+            filepath = self.format_checkpoint_name(
+                epoch, ckpt_name_metrics, ver=version_cnt
+            )
             # this epoch called before
             version_cnt += 1
 
@@ -358,33 +366,54 @@ class ModelCheckpoint(Callback):
 
             if not isinstance(current, torch.Tensor):
                 rank_zero_warn(
-                    f'The metric you returned {current} must be a `torch.Tensor` instance, checkpoint not saved'
-                    f' HINT: what is the value of {self.monitor} in validation_epoch_end()?', RuntimeWarning
+                    f"The metric you returned {current} must be a `torch.Tensor` instance, checkpoint not saved"
+                    f" HINT: what is the value of {self.monitor} in validation_epoch_end()?",
+                    RuntimeWarning,
                 )
                 if current is not None:
                     current = torch.tensor(current)
 
             if current is None:
                 rank_zero_warn(
-                    f'Can save best model only with {self.monitor} available, skipping.', RuntimeWarning
+                    f"Can save best model only with {self.monitor} available, skipping.",
+                    RuntimeWarning,
                 )
             elif self.check_monitor_top_k(current):
                 self._do_check_save(filepath, current, epoch, trainer, pl_module)
-            elif self.verbose > 0:
-                log.info(f'\nEpoch {epoch:05d}: {self.monitor}  was not in top {self.save_top_k}')
+            elif self.verbose:
+                log.info(
+                    f"Epoch {epoch:d}: {self.monitor} was not in top {self.save_top_k}"
+                )
 
         else:
-            if self.verbose > 0:
-                log.info(f'\nEpoch {epoch:05d}: saving model to {filepath}')
+            if self.verbose:
+                log.info(f"Epoch {epoch:d}: saving model to {filepath}")
 
-            assert trainer.global_rank == 0, 'tried to make a checkpoint from non global_rank=0'
+            assert (
+                trainer.global_rank == 0
+            ), "tried to make a checkpoint from non global_rank=0"
             self._save_model(filepath, trainer, pl_module)
 
         if self.save_last:
-            filepath = os.path.join(self.dirpath, self.prefix + ModelCheckpoint.CHECKPOINT_NAME_LAST)
+            filename = self._format_checkpoint_name(
+                self.CHECKPOINT_NAME_LAST, epoch, ckpt_name_metrics, prefix=self.prefix
+            )
+            filepath = os.path.join(self.dirpath, f"{filename}.ckpt")
             self._save_model(filepath, trainer, pl_module)
+            if self.last_model_path and self.last_model_path != filepath:
+                self._del_model(self.last_model_path)
 
-    def _do_check_save(self, filepath, current, epoch, trainer, pl_module):
+    def _is_valid_monitor_key(self, metrics):
+        return self.monitor in metrics or len(metrics) == 0
+
+    def _do_check_save(
+        self,
+        filepath: str,
+        current: torch.Tensor,
+        epoch: int,
+        trainer,
+        pl_module,
+    ):
         # remove kth
 
         del_list = []
@@ -396,32 +425,36 @@ class ModelCheckpoint(Callback):
         self.best_k_models[filepath] = current
         if len(self.best_k_models) == self.save_top_k:
             # monitor dict has reached k elements
-            _op = max if self.mode == 'min' else min
-            self.kth_best_model_path = _op(self.best_k_models,
-                                           key=self.best_k_models.get)
+            _op = max if self.mode == "min" else min
+            self.kth_best_model_path = _op(
+                self.best_k_models, key=self.best_k_models.get
+            )
             self.kth_value = self.best_k_models[self.kth_best_model_path]
 
-        _op = min if self.mode == 'min' else max
+        _op = min if self.mode == "min" else max
         self.best_model_path = _op(self.best_k_models, key=self.best_k_models.get)
         self.best_model_score = self.best_k_models[self.best_model_path]
 
-        if self.verbose > 0:
+        if self.verbose:
             log.info(
-                f'\nEpoch {epoch:05d}: {self.monitor} reached'
-                f' {current:0.5f} (best {self.best_model_score:0.5f}), saving model to'
-                f' {filepath} as top {self.save_top_k}')
+                f"Epoch {epoch:d}: {self.monitor} reached"
+                f" {current:0.5f} (best {self.best_model_score:0.5f}),"
+                f" saving model to {filepath} as top {self.save_top_k}"
+            )
         self._save_model(filepath, trainer, pl_module)
 
         for cur_path in del_list:
             if cur_path != filepath:
                 self._del_model(cur_path)
 
-    def on_save_checkpoint(self, trainer, pl_module):
+    def on_save_checkpoint(
+        self, trainer, pl_module
+    ) -> Dict[str, Any]:
         return {
-            'best_model_score': self.best_model_score,
-            'best_model_path': self.best_model_path,
+            "best_model_score": self.best_model_score,
+            "best_model_path": self.best_model_path,
         }
 
-    def on_load_checkpoint(self, checkpointed_state):
-        self.best_model_score = checkpointed_state['best_model_score']
-        self.best_model_path = checkpointed_state['best_model_path']
+    def on_load_checkpoint(self, checkpointed_state: Dict[str, Any]):
+        self.best_model_score = checkpointed_state["best_model_score"]
+        self.best_model_path = checkpointed_state["best_model_path"]