+++ tests/base/model_test_steps.py
@@ -18,6 +18,8 @@ class TestStepVariations(ABC):
         :param batch:
         :return:
         """
+        self.test_step_called = True
+
         x, y = batch
         x = x.view(x.size(0), -1)
         y_hat = self(x)